Move I2CxHandle out of global memory
This is a bit a trick, since it needs to be global but is only used
in the cold boot case. What we do then is make a global pointer to
the structure and then  zero-init and fill a copy only on the cold
boot path where it's used.

This trade-off makes the 1-time true cold boot a little bit slower
but the long-running warm-boot faster [also trades off a little
code space]
the structure in the cold-boot